requirements.txt
openai==0.27.4
python-redmine==2.4.0
tzdata==2023.3
python-dotenv==1.0.0